Extract static data from TweedleEncoder into TweedleEncoderData
TweedleEncoder: 820 → 623 lines
TweedleEncoderData: 213 lines (new)

Moves all 11 static Map/Set fields and the static initializer block
into a separate data class, following the same pattern as the
ProjectMigrationManager → TextMigrationRegistry extraction.

Continues #506
