+++ .github/hooks/amplihack-hooks.json
@@ -1,0 +1,47 @@
+{
+  "hooks": {
+    "agentStop": [
+      {
+        "bash": "/home/azureuser/src/alice/.github/hooks/stop",
+        "timeoutSec": 30,
+        "type": "command"
+      }
+    ],
+    "postToolUse": [
+      {
+        "bash": "/home/azureuser/src/alice/.github/hooks/post-tool-use",
+        "timeoutSec": 30,
+        "type": "command"
+      }
+    ],
+    "preCompact": [
+      {
+        "bash": "/home/azureuser/src/alice/.github/hooks/pre-compact",
+        "timeoutSec": 30,
+        "type": "command"
+      }
+    ],
+    "preToolUse": [
+      {
+        "bash": "/home/azureuser/src/alice/.github/hooks/pre-tool-use",
+        "timeoutSec": 30,
+        "type": "command"
+      }
+    ],
+    "sessionStart": [
+      {
+        "bash": "/home/azureuser/src/alice/.github/hooks/session-start",
+        "timeoutSec": 30,
+        "type": "command"
+      }
+    ],
+    "userPromptSubmitted": [
+      {
+        "bash": "/home/azureuser/src/alice/.github/hooks/user-prompt-submit",
+        "timeoutSec": 30,
+        "type": "command"
+      }
+    ]
+  },
+  "version": 1
+}
